Revert "[Experiment] Store snapshots in `git lfs` (#643)"
This reverts commit 5aa5937d78dd1ec3b5b067dc173781858c38550e.
