--- .gitattributes
@@ -1,3 +1,0 @@
-# LFS settings
-# If changing, also change in .github/workflows/ci.yml
-vello_tests/snapshots/*.png filter=lfs diff=lfs merge=lfs -text
