--- vello_tests/smoke_snapshots/.gitignore
@@ -1,3 +1,0 @@
-*.oversized.png
-*.new.png
-!.gitignore
