+++ vello_tests/snapshots/.gitignore
@@ -1,3 +1,2 @@
-*.oversized.png
 *.new.png
 !.gitignore
