--- vello_tests/snapshots/fill_types.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:9e8e4396ee504ce60fb2b45a57a58eeff45008b3f78b612df858091752c4e06c
-size 82101
