+++ vello_tests/snapshots/filled_circle.png
