+++ vello_tests/snapshots/filled_square.png
