--- vello_tests/snapshots/funky_paths.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:1eab962256121827ea6632be866c7ea654d6a6be45f8020fa9f5d06a67870988
-size 17989
