+++ vello_tests/snapshots/single_emoji.png
