--- vello_tests/snapshots/splash.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a9146b72893d869c3894dbee87e640904d6d5645b1ca236f3f096aaefb34bcc8
-size 99856
