--- vello_tests/snapshots/stroke_styles.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:172954534a5496a6e2e2883d644c6c073541e0302a216e689ae63cddc799dfae
-size 94334
