--- vello_tests/snapshots/stroke_styles_non_uniform.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:73da696256f7ed8c94a76dc0dfc915a8aca38da9de68c86610c70a3e3e0da18b
-size 88607
