--- vello_tests/snapshots/stroke_styles_skew.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:66591bfd8cea31bbd39f657afd2f25cf30b8b19f8ff4775be416c97e2035572a
-size 97701
