--- vello_tests/snapshots/tricky_strokes.png
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:8d22d397f9a45be14784abe692e0b4fae16e555834876021d74e35d44f872043
-size 34733
